feat: rename distribution pkg & import pkg
