+++ README.md
@@ -1,7 +1,7 @@
 # Docling LangChain integration
 
-[![PyPI version](https://img.shields.io/pypi/v/docling-langchain)](https://pypi.org/project/docling-langchain/)
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/docling-langchain)](https://pypi.org/project/docling-langchain/)
+[![PyPI version](https://img.shields.io/pypi/v/langchain-docling)](https://pypi.org/project/langchain-docling/)
+[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/langchain-docling)](https://pypi.org/project/langchain-docling/)
 [![Poetry](https://img.shields.io/endpoint?url=https://python-poetry.org/badge/v0.json)](https://python-poetry.org/)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Imports: isort](https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336)](https://pycqa.github.io/isort/)
@@ -14,9 +14,9 @@ A [Docling](https://github.com/DS4SD/docling) integration for
 
 ## Installation
 
-Simply install `docling-langchain` from your package manager, e.g. pip:
+Simply install `langchain-docling` from your package manager, e.g. pip:
 ```bash
-pip install docling-langchain
+pip install langchain-docling
 ```
 
 ## Usage
@@ -24,7 +24,7 @@ pip install docling-langchain
 Basic usage looks as follows:
 
 ```python
-from docling_langchain import DoclingLoader
+from langchain_docling import DoclingLoader
 
 FILE_PATH = ["https://arxiv.org/pdf/2408.09869"]  # Docling Technical Report
